+++ relayer/build_processors.go
@@ -10,8 +10,6 @@ import (
 	evmtypes "github.com/palomachain/pigeon/types/paloma/x/evm/types"
 	log "github.com/sirupsen/logrus"
 	"github.com/vizualni/whoops"
-
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 )
 
 func (r *Relayer) buildProcessors(ctx context.Context) ([]chain.Processor, error) {
@@ -69,52 +67,4 @@ func (r *Relayer) processorFactory(chainInfo *evmtypes.ChainInfo) (chain.Process
 		return nil, whoops.Wrap(err, retErr)
 	}
 	return processor, nil
-}
-
-func (r *Relayer) HealthCheck(ctx context.Context) error {
-	chainsInfos, err := r.palomaClient.QueryGetEVMChainInfos(ctx)
-
-	if err != nil {
-		return err
-	}
-
-	val, err := r.palomaClient.GetValidator(ctx)
-	if err != nil {
-		return err
-	}
-
-	if val == nil {
-		// I am sorry, but where *is* your validator?
-		return ErrNotAValidatorAccount
-	}
-
-	isStaking := false
-	if !val.Jailed {
-		if val.Status == stakingtypes.Bonded || val.Status == stakingtypes.Unbonding {
-			isStaking = true
-		}
-	}
-
-	var g whoops.Group
-	for _, chainInfo := range chainsInfos {
-		p, err := r.processorFactory(chainInfo)
-		if err != nil {
-			g.Add(err)
-			continue
-		}
-
-		g.Add(p.HealthCheck(ctx))
-	}
-
-	if !isStaking {
-		// then these errors are only warning
-		log.Warn("validator is not staking. ensure to fix these warning if you wish to stake.")
-		for _, err := range g {
-			log.WithError(err).Warn("blocker for becoming a staking validator. Fix if you wish to stake.")
-		}
-
-		return nil
-	}
-
-	return g.Return()
 }
